Simplify syslog logging
Extra syslog feature support needs to be carefully thought out.  We
kind-of tried to support `ident` parameter, but then there is this in
the man page:

  The argument ident in the call of openlog() is probably stored as-is.
  Thus, if the string it points to is changed, syslog() may start
  prepending the changed string, and if the string it points to ceases
  to exist, the results are undefined.  Most portable is to use a string
  constant.

Not sure if there is a way to guarantee a string will not be moved by
GC.  So far we've probably been just lucky -- maybe we've only used
literal strings and those are put in "static space"?

Also the use of `option` parameter (we used PID) might have been wrong.

For now just use `syslog` function which will automatically use
`openlog` and use program name as `ident`.